优化电流校准模块

+++ SoftWare/STM32G431CBUx/motorlib/driver/currsmp.h
@@ -8,9 +8,9 @@ struct currsmp_data {
 	uint32_t channle_raw_b;
 	uint32_t channle_raw_c;
 
-	float offset_a;
-	float offset_b;
-	float offset_c;
+	uint32_t offset_a;
+	uint32_t offset_b;
+	uint32_t offset_c;
 
 	float factor;
 
@@ -25,6 +25,6 @@ struct currsmp_data {
 
 void currsmp_updata(struct device *dev);
 void currsmp_caribe_offset(struct device *dev);
-void currsmp_update_raw(struct device *currsmp,uint32_t *adc_raw);
+void currsmp_update_raw(struct device *currsmp, uint32_t *adc_raw);
 
 #endif
